Move to util file
